에러처리

+++ src/main/java/com/foreveryone/knowing/service/AdminService.java
@@ -113,6 +113,7 @@ public class AdminService {
     }
 
     public void deleteInquiry(Integer inquiryId) {
+        if (!inquiryRepository.existsById(inquiryId)) throw new NotFoundException("inquiry id not found");
         inquiryRepository.deleteById(inquiryId);
     }
